+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed bug in enclosing functions
 - Parallelized PHR find initial Cut
 - caching subtree size
+- atomic counter for updateAABB
 
 ## [0.0.3] - 2021-08-12
 ### Added